tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/023.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: DOM added script, even later .src </title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <script>console.log('inline script #1');
                var script = testlib.addScript('', { }, document.getElementsByTagName('body')[0], false);
                console.log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        function test() {
                assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2', 'external script #1']);
                t.done();
}
        onload = t.step_func(function(){
                console.log(1);
                script.src='scripts/include-1.js';
                console.log(2);
                script.onload = t.step_func(test);
                console.log(3);
        })
        </script>

</body></html>
